Fix very long time when running Makefile
Running `base.Makefile` takes a significantly longer time with make 4.4.x
compared to make 4.3.x.

This is due to the conditional variable assignment (`?=`) when variables are
exported and receive their value from a call to the `shell` function.

Variables set with `?=` are expanded recursively. Expansion occurs every time
the variable is required. When exported, variables are also expanded each
time any child process is launched. Using `shell` exacerbates the situation
as it is considered to be "slow".

To resolve this issue, this commit replaces `?=` assignments with simply
expanded variable assignments wrapped inside a conditional. As a result,
these variables are assigned only once.

This fix should not change the behavior.

See:

- [Why export variables with shell command in Makefile takes a very long
  time?](https://stackoverflow.com/questions/77363777/why-export-variables-with-shell-command-in-makefile-takes-a-very-long-time)
- [The Two Flavors of Variables](https://www.gnu.org/software/make/manual/html_node/Flavors.html)