Do not drop dashes when converting a ref name to a GitLab-like ref slug
CI jobs on GitLab automatically [define](https://docs.gitlab.com/ci/variables/predefined_variables/)
a variable `CI_COMMIT_REF_SLUG` named after the refs pointing to a commit.

If unset (when run outside of the GitLab CI pipeline), this variable is
explicitely defined, trying to mimic the same behavior as on the CI.

The GitLab's variable do not remove dashes but our implementation did.
The difference is confusing and could introduce subtile bugs.

For instance:

- on GitLab, `this-is-a-slug`
- elsewhere, `thisisaslug`

This change fixes this.